# backend/internal/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Bitik Backend API
  version: 0.1.0
  description: Baseline OpenAPI contract for the Bitik marketplace backend.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/media/upload:
    post:
      summary: Upload seller/admin media
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201": { description: Uploaded media metadata. }
  /api/v1/media/upload/presigned-url:
    post:
      summary: Create presigned seller/admin media upload URL
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Presigned upload URL and reserved media metadata. }
  /api/v1/media/upload/presigned-complete:
    post:
      summary: Complete and validate a presigned media upload
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Validated media metadata. }
  /api/v1/media/files:
    get:
      summary: List current user's media files
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Media files. }
  /api/v1/media/files/{file_id}:
    get:
      summary: Get media file metadata
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: file_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Media file metadata. }
        "404": { description: Media file not found. }
    delete:
      summary: Delete media file
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: file_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Media file deleted. }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalTokenAuth:
      type: apiKey
      in: header
      name: X-Internal-Token
...